Создал класс Апи и вынес туда вызов метода на создание курьера
